+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/balance/request/FetchBalanceCriteria.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concertreservationservice.application.facade.balance.request
+
+data class FetchBalanceCriteria(
+    val token: String,
+)
