+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/concert/request/SearchAvailSeatCriteria.kt
@@ -1,0 +1,10 @@
+package io.hhplus.concertreservationservice.application.facade.concert.request
+
+import java.time.LocalDate
+
+data class SearchAvailSeatCriteria(
+    val token: String,
+    val concertId: Long,
+    val scheduleId: Long,
+    val date: LocalDate,
+)
