+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/concert/request/SeatReserveCriteria.kt
@@ -1,0 +1,8 @@
+package io.hhplus.concertreservationservice.application.facade.concert.request
+
+data class SeatReserveCriteria(
+    val concertId: Long,
+    val scheduleId: Long,
+    val seatNo: Int,
+    val token: String,
+)
