+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/payment/request/ProcessPaymentCriteria.kt
@@ -1,0 +1,7 @@
+package io.hhplus.concertreservationservice.application.facade.payment.request
+
+class ProcessPaymentCriteria(
+    val token: String,
+    val reservationId: Long,
+    val amount: Long,
+)
