+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/token/request/CreateReservationTokenCriteria.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concertreservationservice.application.facade.token.request
+
+data class CreateReservationTokenCriteria(
+    val userId: Long,
+)
