+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/token/request/ReservationTokenStatusCriteria.kt
@@ -1,0 +1,5 @@
+package io.hhplus.concertreservationservice.application.facade.token.request
+
+data class ReservationTokenStatusCriteria(
+    val token: String,
+)
