+++ src/main/kotlin/io/hhplus/concertreservationservice/application/facade/token/response/CreateReservationTokenResult.kt
@@ -1,0 +1,14 @@
+package io.hhplus.concertreservationservice.application.facade.token.response
+
+import io.hhplus.concertreservationservice.presentation.controller.token.response.ReservationTokenCreateResponse
+
+data class CreateReservationTokenResult(
+    val userId: Long,
+    val token: String,
+)
+
+fun CreateReservationTokenResult.toResponse(): ReservationTokenCreateResponse {
+    return ReservationTokenCreateResponse(
+        token = this.token,
+    )
+}
