--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/balance/request/FetchBalanceCriteria.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.balance.request
-
-data class FetchBalanceCriteria(
-    val token: String,
-)
