--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/balance/response/ChargeBalanceResult.kt
@@ -1,16 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.balance.response
-
-import io.hhplus.concertreservationservice.domain.balance.Money
-import io.hhplus.concertreservationservice.presentation.controller.balance.response.BalanceChargeResponse
-
-data class ChargeBalanceResult(
-    val userId: Long,
-    val userName: String,
-    val balance: Money,
-)
-
-fun ChargeBalanceResult.toResponse(): BalanceChargeResponse {
-    return BalanceChargeResponse(
-        balance = balance.amount,
-    )
-}
