--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/balance/response/FetchBalanceResult.kt
@@ -1,16 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.balance.response
-
-import io.hhplus.concertreservationservice.domain.balance.Money
-import io.hhplus.concertreservationservice.presentation.controller.balance.response.BalanceFetchResponse
-
-data class FetchBalanceResult(
-    val userId: Long,
-    val userName: String,
-    val balance: Money,
-)
-
-fun FetchBalanceResult.toResponse(): BalanceFetchResponse {
-    return BalanceFetchResponse(
-        balance = balance.amount,
-    )
-}
