--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/concert/request/SeatReserveCriteria.kt
@@ -1,8 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.concert.request
-
-data class SeatReserveCriteria(
-    val concertId: Long,
-    val scheduleId: Long,
-    val seatNo: Int,
-    val token: String,
-)
