--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/payment/request/ProcessPaymentCriteria.kt
@@ -1,7 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.payment.request
-
-class ProcessPaymentCriteria(
-    val token: String,
-    val reservationId: Long,
-    val amount: Long,
-)
