--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/payment/response/ProcessPaymentResult.kt
@@ -1,15 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.payment.response
-
-import io.hhplus.concertreservationservice.domain.payment.PaymentStatus
-import io.hhplus.concertreservationservice.presentation.controller.payment.response.PaymentResponse
-
-class ProcessPaymentResult(
-    val paymentId: Long,
-    val status: PaymentStatus,
-)
-
-fun ProcessPaymentResult.toResponse(): PaymentResponse {
-    return PaymentResponse(
-        status = this.status,
-    )
-}
