--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/token/request/CreateReservationTokenCriteria.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.token.request
-
-data class CreateReservationTokenCriteria(
-    val userId: Long,
-)
