--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/token/request/ReservationTokenStatusCriteria.kt
@@ -1,5 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.token.request
-
-data class ReservationTokenStatusCriteria(
-    val token: String,
-)
