--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/token/response/CreateReservationTokenResult.kt
@@ -1,14 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.token.response
-
-import io.hhplus.concertreservationservice.presentation.controller.token.response.ReservationTokenCreateResponse
-
-data class CreateReservationTokenResult(
-    val userId: Long,
-    val token: String,
-)
-
-fun CreateReservationTokenResult.toResponse(): ReservationTokenCreateResponse {
-    return ReservationTokenCreateResponse(
-        token = this.token,
-    )
-}
