--- src/main/kotlin/io/hhplus/concertreservationservice/application/usecase/token/response/ReservationTokenStatusResult.kt
@@ -1,17 +1,0 @@
-package io.hhplus.concertreservationservice.application.usecase.token.response
-
-import io.hhplus.concertreservationservice.domain.token.TokenStatus
-import io.hhplus.concertreservationservice.presentation.controller.token.response.ReservationTokenStatusResponse
-
-data class ReservationTokenStatusResult(
-    val token: String,
-    val status: TokenStatus,
-    val userId: Long,
-)
-
-fun ReservationTokenStatusResult.toResponse(): ReservationTokenStatusResponse {
-    return ReservationTokenStatusResponse(
-        status = this.status,
-        userId = this.userId,
-    )
-}
