+++ src/main/kotlin/io/hhplus/concertreservationservice/domain/concert/service/response/CreateReservedSeatInfo.kt
@@ -1,6 +1,6 @@
 package io.hhplus.concertreservationservice.domain.concert.service.response
 
-import io.hhplus.concertreservationservice.application.usecase.concert.response.SeatReserveResult
+import io.hhplus.concertreservationservice.application.facade.concert.response.SeatReserveResult
 
 data class CreateReservedSeatInfo(
     val reservationId: Long,
