+++ src/main/kotlin/io/hhplus/concertreservationservice/domain/payment/service/response/ProcessPaymentInfo.kt
@@ -1,6 +1,6 @@
 package io.hhplus.concertreservationservice.domain.payment.service.response
 
-import io.hhplus.concertreservationservice.application.usecase.payment.response.ProcessPaymentResult
+import io.hhplus.concertreservationservice.application.facade.payment.response.ProcessPaymentResult
 import io.hhplus.concertreservationservice.domain.payment.PaymentStatus
 
 class ProcessPaymentInfo(
