+++ src/main/kotlin/io/hhplus/concertreservationservice/domain/token/service/response/TokenStatusInfo.kt
@@ -1,6 +1,6 @@
 package io.hhplus.concertreservationservice.domain.token.service.response
 
-import io.hhplus.concertreservationservice.application.usecase.token.response.ReservationTokenStatusResult
+import io.hhplus.concertreservationservice.application.facade.token.response.ReservationTokenStatusResult
 import io.hhplus.concertreservationservice.domain.token.TokenStatus
 import java.time.LocalDateTime
